+++ client/client_test.go
@@ -1869,6 +1869,110 @@ func TestOpenFgaClient(t *testing.T) {
 		}
 	})
 
+	t.Run("ListUsers", func(t *testing.T) {
+		test := TestDefinition{
+			Name: "ListUsers",
+			// A real API would not return all these for the filter provided, these are just for test purposes
+			JsonResponse:   `{"excluded_users":null,"users":[{"object":{"id":"81684243-9356-4421-8fbf-a4f8d36aa31b","type":"user"}},{"userset":{"id":"fga","relation":"member","type":"team"}},{"wildcard":{"type":"user"}}]}`,
+			ResponseStatus: http.StatusOK,
+			Method:         http.MethodPost,
+			RequestPath:    "list-users",
+		}
+
+		requestBody := ClientListUsersRequest{
+			Object: openfga.FgaObject{
+				Type: "document",
+				Id:   "roadmap",
+			},
+			Relation: "can_read",
+			// API does not allow sending multiple filters, done here for testing purposes
+			UserFilters: []openfga.UserTypeFilter{{
+				Type: "user",
+			}, {
+				Type:     "team",
+				Relation: openfga.PtrString("member"),
+			}},
+			ContextualTuples: []ClientContextualTupleKey{{
+				User:     "user:81684243-9356-4421-8fbf-a4f8d36aa31b",
+				Relation: "editor",
+				Object:   "folder:product",
+			}, {
+				User:     "folder:product",
+				Relation: "parent",
+				Object:   "document:roadmap",
+			}},
+			Context: &map[string]interface{}{"ViewCount": 100},
+		}
+		options := ClientListUsersOptions{
+			AuthorizationModelId: openfga.PtrString("01GAHCE4YVKPQEKZQHT2R89MQV"),
+		}
+
+		var expectedResponse openfga.ListUsersResponse
+		if err := json.Unmarshal([]byte(test.JsonResponse), &expectedResponse); err != nil {
+			t.Fatalf("%v", err)
+		}
+
+		httpmock.Activate()
+		defer httpmock.DeactivateAndReset()
+		httpmock.RegisterResponder(test.Method, fmt.Sprintf("%s/stores/%s/%s", fgaClient.GetConfig().ApiUrl, fgaClient.GetConfig().StoreId, test.RequestPath),
+			func(req *http.Request) (*http.Response, error) {
+				resp, err := httpmock.NewJsonResponse(test.ResponseStatus, expectedResponse)
+				if err != nil {
+					return httpmock.NewStringResponse(http.StatusInternalServerError, ""), nil
+				}
+				return resp, nil
+			},
+		)
+
+		got, err := fgaClient.ListUsers(context.Background()).
+			Body(requestBody).
+			Options(options).
+			Execute()
+
+		_, err = got.MarshalJSON()
+		if err != nil {
+			t.Fatalf("%v", err)
+		}
+
+		if len(got.Users) != len(expectedResponse.Users) {
+			t.Fatalf("OpenFga%v().Execute() = %v, want %v", test.Name, got.GetUsers(), expectedResponse.GetUsers())
+		}
+
+		if got.Users[0].GetObject().Type != expectedResponse.Users[0].GetObject().Type || got.Users[0].GetObject().Id != expectedResponse.Users[0].GetObject().Id {
+			t.Fatalf("OpenFga%v().Execute() = %v, want %v (%v)", test.Name, got.Users[0], expectedResponse.Users[0], "object: { type: \"user\", id: \"81684243-9356-4421-8fbf-a4f8d36aa31b\" }")
+		}
+
+		if got.Users[1].GetUserset().Type != expectedResponse.Users[1].GetUserset().Type || got.Users[1].GetUserset().Id != expectedResponse.Users[1].GetUserset().Id || got.Users[1].GetUserset().Relation != expectedResponse.Users[1].GetUserset().Relation {
+			t.Fatalf("OpenFga%v().Execute() = %v, want %v (%v)", test.Name, got.Users[1], expectedResponse.Users[1], "wildcard: { type: \"team\", id: \"fga\", relation: \"member\" }")
+		}
+
+		if got.Users[2].GetWildcard().Type != expectedResponse.Users[2].GetWildcard().Type {
+			t.Fatalf("OpenFga%v().Execute() = %v, want %v (%v)", test.Name, got.Users[2], expectedResponse.Users[2], "wildcard: { type: \"user\" }")
+		}
+
+		if len(got.ExcludedUsers) != len(expectedResponse.ExcludedUsers) {
+			t.Fatalf("OpenFga%v().Execute() = %v, want %v", test.Name, got.GetExcludedUsers(), expectedResponse.GetExcludedUsers())
+		}
+
+		// ListUsers without options should work
+		_, err = fgaClient.ListUsers(context.Background()).Body(requestBody).Execute()
+		if err != nil {
+			t.Fatalf("%v", err)
+		}
+
+		// Invalid auth model id should result in error
+		badOptions := ClientListUsersOptions{
+			AuthorizationModelId: openfga.PtrString("INVALID"),
+		}
+		_, err = fgaClient.ListUsers(context.Background()).
+			Body(requestBody).
+			Options(badOptions).
+			Execute()
+		if err == nil {
+			t.Fatalf("Expect error with invalid auth model id but there is none")
+		}
+	})
+
 	/* Assertions */
 	t.Run("ReadAssertions", func(t *testing.T) {
 		modelId := "01GAHCE4YVKPQEKZQHT2R89MQV"